Transfer OnPositionChangeEventArgs to it's own file
