+++ Data/OnPositionChangeEventArgs.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace TPW.Data;
+
+public class OnPositionChangeEventArgs : EventArgs
+{
+   public readonly IBall Ball;
+
+   public OnPositionChangeEventArgs(IBall ball)
+   {
+      this.Ball = ball;
+   }
+}
