+++ Logic/BallsLogicLayerAbstractApi.cs
@@ -4,16 +4,6 @@ using TPW.Data;
 
 namespace TPW.Logic;
 
-public class OnPositionChangeEventArgs : EventArgs
-{
-	public readonly ILogicBall Ball;
-
-	public OnPositionChangeEventArgs(ILogicBall ball)
-	{
-		this.Ball = ball;
-	}
-}
-
 public abstract class BallsLogicLayerAbstractApi
 {
 	public event EventHandler<OnPositionChangeEventArgs>? PositionChange;
