+++ Logic/OnPositionChangeEventArgs.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+namespace TPW.Logic;
+
+public class OnPositionChangeEventArgs : EventArgs
+{
+   public readonly ILogicBall Ball;
+
+   public OnPositionChangeEventArgs(ILogicBall ball)
+   {
+      this.Ball = ball;
+   }
+}
